Initialize lock_service
